refactor: 사용자 입력에 관한 클래스와 enum의 패키지와 코드 수정
