+++ src/main/java/bridge/command/enums/GameCommand.java
@@ -1,4 +1,4 @@
-package bridge.option;
+package bridge.command.enums;
 
 import java.util.Arrays;
