+++ src/main/java/bridge/command/enums/PlayerMove.java
@@ -1,4 +1,4 @@
-package bridge.player;
+package bridge.command.enums;
 
 import java.util.Arrays;
