+++ src/test/java/bridge/OptionTest.java
@@ -3,9 +3,9 @@ package bridge;
 import static bridge.config.ExceptionMessage.ERROR;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import bridge.option.Command;
-import bridge.option.Move;
-import bridge.option.Option;
+import bridge.command.Command;
+import bridge.command.Move;
+import bridge.command.Option;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
